Refactor: Move hashwalker in folder
